storage: clamp epoch-ms SQL bounds instead of casting u64 raw (#457)
Time bounds reach the SQL backends as `u64` but every `*_at_ms` column is
signed 64-bit, so a raw `as i64` cast of a bound above `i64::MAX` wraps to a
negative value and inverts the comparison. A host passing `u64::MAX` got the
complement of the in-memory answer: attachment cutoffs stopped matching aged
intents, trigger occurrence listings returned nothing where the in-memory
`TriggerOccurrenceFilter::matches` returned everything, and the process-list
time predicates pushed down the opposite range.

Every cutoff/bound now goes through one crate-local `clamp_epoch_ms` helper
per SQL backend, which saturates at `i64::MAX` — exact for every timestamp
below `i64::MAX` — so the class cannot drift back one call site at a time.
The `created_at_end_ms` site that hand-rolled its own out-of-range guard is
normalized onto the same helper.

Cross-backend laws cover it: the attachment large-cutoff conformance law, a
new trigger-store law asserting the occurrence pushdown equals the Rust
predicate for bounds at and above `i64::MAX`, and additional huge-bound
parity cases in the process-list filter law. Reverting either store's clamp
reproduces the divergence on sqlite and postgres.

Release-Notes: Fixed: SQL store backends no longer truncate very large u64 time cutoffs to -1; attachment, trigger, and process-list time filters now behave identically across backends.

+++ crates/lash-core/src/testing/conformance/runtime_persistence_state_machine/attachment_conservation.rs
@@ -7,7 +7,7 @@
 
 use super::*;
 
-const RECONCILE_SQL_SAFE_MAX: u64 = i64::MAX as u64;
+const RECONCILE_SQL_SAFE_MAX: u64 = u64::MAX;
 
 /// Production handles required by the runtime-persistence state machine.
 ///
